opencdk: small fixed to reduce warnings

+++ lib/opencdk/kbnode.c
@@ -522,7 +522,7 @@ cdk_kbnode_write_to_mem(cdk_kbnode_t node, byte * buf, size_t * r_nbytes)
 		cdk_stream_close(s);
 		return 0;
 	}
-	if (*r_nbytes < len) {
+	if (*r_nbytes < (size_t)len) {
 		*r_nbytes = len;
 		rc = CDK_Too_Short;
 	}
